koshelek-mvp/pyproject.toml
[tool.poetry]
name = "koshelek-mvp"
version = "0.1.0"
description = ""
authors = ["Beknazar Abdyldaev <abdyldaevgroup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.2"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
django-environ = "^0.9.0"
requests = "^2.28.2"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.4"
psycopg2-binary = "^2.9.5"
celery = "^5.2.7"
redis = "^4.4.2"
gunicorn = "^20.1.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"



[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.11.4"
django-stubs = "^1.14.0"
djangorestframework-stubs = "^1.8.0"
drf-yasg-stubs = "^0.1.4"


[tool.isort]
profile = "black"
skip_glob = ['*/migrations/*', "*/asgi.py"]
skip = [".gitignore", ".dockerignore", "env", ".venv"]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
exclude = "(migrations|settings|tests|manage.py)"


allow_redefinition = true

follow_imports = "silent"

check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true

warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true

strict_optional = true
strict_equality = true

[tool.django-stubs]
django_settings_module = "core.settings"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
